# contract/openapi.yaml
openapi: 3.1.0
info:
  title: 3D Review API — contract-v0.1 (DRAFT)
  version: 0.1.0-draft
  description: Projected from cookbook 04 @ 0f4413768c4365054372d917619311d7c2a00e96 by Fable. 84 capabilities; every capability
    is also reachable via MCP docs/read/write/danger over the same handlers (prd/18-D). Paths marked x-path-inferred were
    elided in 04 and need Lane A confirmation. SUPPRESSED is ok:true. Danger twins are never GET.
servers:
- url: https://{host}/
  variables:
    host:
      default: TBD-OF-2
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: Cloudflare email-code login → HttpOnly cookie (18-C)
    bearer:
      type: http
      scheme: bearer
      description: delegated agent bearer — acts as a user, never as a role (18-C / MCP-REQ-007)
    participant:
      type: http
      scheme: bearer
      description: participant token from redeem_code / open_link, bound to one assessment survey
  schemas:
...
    Receipt:
      type: object
      required:
      - id
      - actor
      - scope
      - class
      - trace_id
      - at
      properties:
        id:
          type: string
        actor:
          type: string
        scope:
          type: object
          properties:
            type:
              type: string
              enum:
              - workspace
              - project
              - assessment
              - survey
              - platform
            id:
              type: string
        class:
          type: string
          enum:
          - read
          - write.reversible
          - write.effect
          - write.dangerous
        undo_token:
          type: string
          description: present only when the row declares a true inverse
        inverse:
          type: string
          description: capability id or 'none'
        compensating_control:
          type: string
        trace_id:
          type: string
        at:
          type: string
          format: date-time
    Envelope:
      type: object
      required:
      - ok
      - capability
      - result
      properties:
        ok:
          const: true
        capability:
          type: string
        result:
          type: object
          additionalProperties: true
          properties:
            suppressed:
              type: boolean
              description: small-cell suppression applied; this is a success state
        receipt:
          $ref: '#/components/schemas/Receipt'
    ErrorEnvelope:
      type: object
      required:
      - ok
      - error
      properties:
        ok:
          const: false
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              enum:
              - NOT_AUTHENTICATED
              - NOT_AUTHORIZED_AT_SCOPE
              - WRONG_TOOL_FOR_CLASS
              - CONFIRM_REQUIRED
              - CONFIRM_EXPIRED
              - INVALID_PARAMS
              - NOT_FOUND_OR_NOT_VISIBLE
              - STAGE_CONFLICT
              - RESERVED_NOT_BUILT
              - NO_INVERSE
              - RATE_LIMITED
            message:
              type: string
            hint:
              type: string
            docs:
              type: string
              description: capability id or topic
...
paths:
...
  /v2/auth/session:
    post:
      operationId: cap.auth.consume_link
      summary: 'cap.auth.consume_link — write.reversible — roles: V'
      x-capability: cap.auth.consume_link
      x-class: write.reversible
      x-tool: write
      x-slice: v2.0-bcs
      x-inverse:
        kind: none
        compensating_control: cap.auth.logout (compensating)
      x-roles: V
      x-path-inferred: false
      tags:
      - A. Entry, identity, session
      responses:
        '200':
          description: ok envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        4XX:
          description: error envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - session: []
      - bearer: []
      - participant: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: params — shape owed per capability (A1 open item)
    delete:
      operationId: cap.auth.logout
      summary: 'cap.auth.logout — write.reversible — roles: any'
      x-capability: cap.auth.logout
      x-class: write.reversible
      x-tool: write
      x-slice: v2.0-bcs
      x-inverse:
        kind: none
        compensating_control: sign in again
      x-roles: any
      x-path-inferred: false
      tags:
      - A. Entry, identity, session
      responses:
        '200':
          description: ok envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        4XX:
          description: error envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - session: []
      - bearer: []
      - participant: []
...
  /v2/participate/receipt:
    get:
      operationId: cap.response.receipt
      summary: 'cap.response.receipt — read — roles: P'
      x-capability: cap.response.receipt
      x-shared-link-lifetime: "Private to authenticated respondent/survey; shared link/session must remain live. Closure/archive does not delete or expose receipts; revocation/expiry denies access."
      x-class: read
      x-tool: read
      x-slice: v2.0-bcs
      x-inverse:
        kind: n/a
      x-roles: P
      x-path-inferred: false
      tags:
      - G. Participation
      responses:
        '200':
          description: ok envelope
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Envelope'
                - {"type":"object","properties":{"result":{"type":"object","properties":{"submitted":{"type":"boolean"},"response_id":{"type":["string","null"]},"submitted_at":{"type":["string","null"],"format":"date-time"},"template":{"oneOf":[{"type":"null"},{"type":"object","properties":{"id":{"type":"string"},"version":{"type":"integer"}},"required":["id","version"],"additionalProperties":false}]}},"required":["submitted","response_id","submitted_at","template"],"additionalProperties":false}},"required":["result"],"additionalProperties":true}
        4XX:
          description: error envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - participant: []
...
  /v2/projects/{id}/rollup:
    get:
      operationId: cap.rollup.project
      summary: 'cap.rollup.project — read — roles: Vw+'
      x-capability: cap.rollup.project
      x-class: read
      x-tool: read
      x-slice: v2.1-oct
      x-inverse:
        kind: n/a
      x-roles: Vw+
      x-path-inferred: true
      tags:
      - H. Results, reports, recommendations (all rows in the whole-product denominator; summary is `v2.0-bcs`, the rest `v2.1-oct`)
      responses:
        '501':
          description: RESERVED_NOT_BUILT — documented, docs pointer; v2.1-oct
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      x-status: RESERVED_NOT_BUILT
      security:
      - session: []
      - bearer: []
      - participant: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /v2/openapi.yaml:
    get:
      operationId: cap.docs.openapi
      summary: 'cap.docs.openapi — read — roles: any'
      x-capability: cap.docs.openapi
      x-class: read
      x-tool: read
      x-slice: v2.0-bcs
      x-inverse:
        kind: n/a
      x-roles: any
      x-path-inferred: false
      tags:
      - J. Ops, telemetry, xray, docs, feedback
      responses:
        '200':
          description: ok envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        4XX:
          description: error envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security: []
  /v2/capabilities.json:
    get:
      operationId: cap.docs.capabilities
      summary: 'cap.docs.capabilities — read — roles: any'
      x-capability: cap.docs.capabilities
      x-class: read
      x-tool: read
      x-slice: v2.0-bcs
      x-inverse:
        kind: n/a
      x-roles: any
      x-path-inferred: false
      tags:
      - J. Ops, telemetry, xray, docs, feedback
      responses:
        '200':
          description: ok envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        4XX:
          description: error envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security: []
...